Modified checkout and refactored parser tests
