+++ src/test/java/seedu/duke/parser/ExitParserTest.java
@@ -1,0 +1,30 @@
+package seedu.duke.parser;
+
+import org.junit.jupiter.api.Test;
+import seedu.duke.commands.ExitCommand;
+import seedu.duke.exceptions.InvalidCommandException;
+import seedu.duke.exceptions.InvalidIdException;
+import seedu.duke.exceptions.InvalidIntegerException;
+import seedu.duke.exceptions.InvalidNameFormatException;
+import seedu.duke.exceptions.InvalidPhoneNumberException;
+import seedu.duke.exceptions.NoArgumentPassedException;
+import seedu.duke.exceptions.PersonNotFoundException;
+import seedu.duke.exceptions.StorageOperationException;
+import seedu.duke.exceptions.WrongFlagException;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class ExitParserTest {
+
+    Parser parser = new Parser();
+
+    @Test
+    public void parseExit_testCommand() throws NoArgumentPassedException,
+            WrongFlagException, InvalidCommandException, InvalidIdException,
+            InvalidPhoneNumberException, InvalidNameFormatException,
+            StorageOperationException, InvalidIntegerException, PersonNotFoundException {
+
+        ExitCommand exitCommand = (ExitCommand) parser.parseCommand("exit");
+        assertEquals("exit", exitCommand.COMMAND);
+    }
+}
